moved Main to GameLauncher
